pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-translate"
version = "1.0.0"
description = "Zero-config AI localization for developers. One command. Every platform."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "Happy-Cyber", email = "dhamasaniyahappy3931@gmail.com" },
]
keywords = [
    "translation", "localization", "i18n", "l10n",
    "ai", "django", "flask", "fastapi", "flutter", "android", "ios",
    "gettext", "arb", "strings", "xml",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Internationalization",
    "Topic :: Software Development :: Localization",
]
dependencies = [
    "requests>=2.28",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "polib>=1.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.30"]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.5"]
mistral = ["mistralai>=1.0"]
all = [
    "anthropic>=0.30",
    "openai>=1.0",
    "google-generativeai>=0.5",
    "mistralai>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
ai-translate = "ai_translate.cli.main:main"

[project.urls]
Homepage = "https://github.com/Happy-cyber/ai-translate"
Repository = "https://github.com/Happy-cyber/ai-translate"
Issues = "https://github.com/Happy-cyber/ai-translate/issues"

[tool.setuptools.packages.find]
include = ["ai_translate*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
